add filtering, sorting

+++ Controllers/CommentController.cs
@@ -31,9 +31,12 @@ namespace api.Controllers
         }
 
         [HttpGet]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
 
             var comment = await _commentRepo.GetByIdAsync(id);
             if (comment == null)
@@ -44,8 +47,12 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto commentDto){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            
             if(!await _stockRepo.StockExists(stockId)){
                 return BadRequest("Stock does not exist");
             }
@@ -55,8 +62,12 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetById), new {id = commentModel.Id}, commentModel.ToCommentDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto commentDto){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            
             var commentModel = await _commentRepo.UpdateAsync(id,commentDto.ToCommentFromUpdate());
             
             if(commentModel == null){
@@ -66,8 +77,12 @@ namespace api.Controllers
             return Ok(commentModel.ToCommentDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            
             var comment = await _commentRepo.DeleteAsync(id);
             if(comment == null){
                 return NotFound();
